refactor(modelHandlers): move openRouterReasoning into its provider dir; add SDK-readiness checkpoint
The OpenRouter reasoning-detail helper imported `@openrouter/sdk/models` while
living in `modelHandlers/utils/`, the tree the README designates for stateless
*cross-provider* helpers. Both its consumers are in `openrouter/`. Relocate it
to `openrouter/openRouterReasoning.ts` so the shared `utils/` tree stops
carrying a provider-SDK type import — tightening the shared/provider boundary
and shrinking the provider-type-leak hazard surface the frozen SDK declaration
graph guards against. The genuinely cross-provider engine it builds on
(`reasoningDetailsText.ts`, also used by the OpenAI-compatible MiniMax handler)
correctly stays in `utils/`. Behavior-preserving; no baseline moves.

Also adds the 2026-08-15 Agent-SDK readiness re-verification checkpoint: a
four-area deep audit (core, model handler, logger, surface) that re-confirms the
standing verdict — well-aligned, no structural refactor, no redundant
abstraction to remove — re-verifies every prior-pass metric at HEAD, catalogues
the small tracked opportunities the deep audit surfaced (residual ambient-ALS
read in the cycle flow; two parallel module-logger factories; process-global log
sink state as the multi-instance-embedding obstacle; a dead redaction-options
branch; the AgentRunStream async-queue extraction), and confirms subagent
boundaries are already drawn.
